# swagger.yml
openapi: 3.0.3
info:
  title: Systeme de Pointage API
  description: Documentation des endpoints disponibles pour l'authentification, le pointage, le pilotage des cours et la gestion academique.
  version: 1.1.0
servers:
  - url: http://127.0.0.1:5000
    description: Serveur local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    kioskKeyAuth:
      type: apiKey
      in: header
      name: X-Attendance-Kiosk-Key
  schemas:
...
    CourseCancelRequest:
      type: object
      required: [motif]
      properties:
        motif:
          type: string
    CourseRescheduleRequest:
      type: object
      required: [date, heure_debut, heure_fin]
      properties:
        date:
          type: string
          example: "2026-07-12"
          description: "Champ canonique. Alias accepte: new_date"
        heure_debut:
          type: string
          example: "15:00:00"
          description: "Champ canonique. Alias accepte: new_start_time"
        heure_fin:
          type: string
          example: "17:00:00"
          description: "Champ canonique. Alias accepte: new_end_time"
        motif:
          type: string
          description: "Champ canonique. Alias accepte: reason"
        salle_id:
          type: integer
          description: "Champ canonique. Alias accepte: room_id"
        new_date:
          type: string
        new_start_time:
          type: string
        new_end_time:
          type: string
        reason:
          type: string
        room_id:
          type: integer
    AttendanceRegularizeRequest:
      type: object
      required: [etudiant_id, statut]
      properties:
        etudiant_id:
          type: integer
          description: "Champ canonique. Alias accepte: student_id"
        statut:
          type: string
          enum: [present, retard, absent, absence_justifiee]
          description: "Champ canonique. Alias accepte: status"
        justificatif:
          type: string
          description: "Champ canonique. Alias accepte: reason"
        pointage_time:
          type: string
          format: date-time
          description: "Horodatage manuel optionnel. Alias accepte: attendance_time"
        student_id:
          type: integer
        status:
          type: string
          enum: [present, retard, absent, absence_justifiee]
        reason:
          type: string
        attendance_time:
          type: string
          format: date-time
...
paths:
...
  /api/v1/professeur/courses:
    get:
      tags: [Professeur]
      summary: Lister les cours accessibles
      security:
        - bearerAuth: []
      parameters:
        - name: statut
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Liste des cours
  /api/v1/professeur/courses/{cours_id}:
    get:
      tags: [Professeur]
      summary: Consulter le detail d'un cours
      security:
        - bearerAuth: []
      parameters:
        - name: cours_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detail du cours
        '403':
          description: Acces refuse
  /api/v1/professeur/courses/{cours_id}/qr:
    get:
      tags: [Professeur]
      summary: Generer le QR d'un cours
      security:
        - bearerAuth: []
      parameters:
        - name: cours_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: QR genere
        '403':
          description: Acces refuse
  /api/v1/professeur/courses/{cours_id}/attendance:
    get:
      tags: [Professeur]
      summary: Voir la feuille de presence d'un cours
      security:
        - bearerAuth: []
      parameters:
        - name: cours_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Etat de presence retourne
  /api/v1/professeur/courses/{cours_id}/attendance/stream:
    get:
      tags: [Professeur]
      summary: Suivre la feuille de presence d'un cours en temps reel (Server-Sent Events)
      description: >-
        Alternative push au polling de GET .../attendance. Reponse
        text/event-stream : un premier evenement `data:` est envoye immediatement
        avec le payload courant (identique a GET .../attendance), puis un nouvel
        evenement des qu'un pointage est enregistre (borne, self-service deprecie)
        ou regularise sur ce cours. Un commentaire `: keep-alive` est envoye toutes
        les 15s en l'absence d'evenement, et la connexion se ferme au bout de 30
        minutes (le navigateur EventSource se reconnecte automatiquement). Limitation
        connue : la diffusion se fait en memoire, par process ; en deploiement
        multi-worker, seuls les abonnes connectes au meme worker que le pointage
        recoivent l'evenement (voir app/services/attendance_events.py).
      security:
        - bearerAuth: []
      parameters:
        - name: cours_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Flux text/event-stream ouvert
          content:
            text/event-stream:
              schema:
                type: string
        '403':
          description: Cours d'un autre professeur
  /api/v1/professeur/courses/{cours_id}/cancel:
    post:
      tags: [Professeur]
      summary: Annuler un cours
      security:
        - bearerAuth: []
      parameters:
        - name: cours_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCancelRequest'
      responses:
        '200':
          description: Cours annule
  /api/v1/professeur/courses/{cours_id}/reschedule:
    post:
      tags: [Professeur]
      summary: Reporter un cours
      security:
        - bearerAuth: []
      parameters:
        - name: cours_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseRescheduleRequest'
      responses:
        '200':
          description: Cours reporte
  /api/v1/professeur/courses/{cours_id}/attendance/regularize:
    post:
      tags: [Professeur]
      summary: Regulariser manuellement un pointage
      security:
        - bearerAuth: []
      parameters:
        - name: cours_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceRegularizeRequest'
      responses:
        '200':
          description: Pointage regularise
        '400':
          description: Donnees invalides
        '409':
          description: Conflit sur le pointage
  /api/v1/professeur/notifications:
    get:
      tags: [Professeur]
      summary: Lister les notifications du professeur ou du responsable dans la vue enseignement
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Notifications retournees
  /api/v1/professeur/notifications/{notification_id}/read:
    patch:
      tags: [Professeur]
      summary: Marquer une notification professeur comme lue
      security:
        - bearerAuth: []
      parameters:
        - name: notification_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Notification mise a jour
        '404':
          description: Notification introuvable
...